Make c10::TempFile non-copyable but movable (#57308)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/57308

This diff makes `c10::TempFile` non-copyable but movable. `torch_shm_manager` was previously dependent upon some hidden behavior that was a result of copying `TempFile`s, which is also being made more explicit now that they can be moved but not copied.

Context:

`c10::TempFile` is currently copyable, which leads to surprising behavior. A seemingly valid `TempFile` may in fact be invalid if the original it was copied from has already been destroyed, resulting in the file descriptor to be closed and the filename being unlinked without the user knowing about it.

**In fact, both `c10::try_make_tempfile` and `c10::make_tempfile` cause copies of `TempFile` to be made**, which can easily be verified by explicitly deleting the copy constructor of `TempFile` and attempting to compile. This means that in practice, users of these functions are getting temporary files that have already been closed and unlinked.

This copying of `TempFile` is particularly interesting in the case of `torch_shm_manager`, which uses `try_make_tempfile` to generate the name of a Unix domain socket to communicate with clients. In order for `bind()` on the socket name to be successful, a file with that same name must not be linked in the filesystem, or `EADDRINUSE` will result. Happily, beacuse `try_make_tempfile` previously created a copy of the `TempFile` while destroying the original, `torch_shm_manager` did not encounter this. With this change, howevrer, `torch_shm_manager` must now explicitly destroy the `TempFile` before attempting to `bind()`. Unfortunately, this exposes a race condition--**other code can re-generate the same-named temporary file after the one created by `torch_shm_manager` is explicitly unlinked but before `torch_shm_manager` binds it to the server socket.** To be clear: this race condition already existed before this diff, but this makes things more explicit. The real fix will be in a follow-up change.

Reviewed By: ejguan

Differential Revision: D28047915

fbshipit-source-id: e8a1b6bb50419fe65620cfecdb67c566a4cf9056

+++ c10/util/tempfile.h
@@ -55,17 +55,34 @@ inline std::vector<char> make_filename(std::string name_prefix) {
 
 struct TempFile {
 #if !defined(_WIN32)
+  TempFile() : fd(-1) {}
   TempFile(std::string name, int fd) : fd(fd), name(std::move(name)) {}
-
-  ~TempFile() {
-    unlink(name.c_str());
-    close(fd);
+  TempFile(const TempFile&) = delete;
+  TempFile(TempFile&& other) noexcept : fd(other.fd), name(std::move(other.name)) {
+    other.fd = -1;
+    other.name.clear();
   }
 
-  const int fd;
+  TempFile& operator=(const TempFile&) = delete;
+  TempFile& operator=(TempFile&& other) {
+    fd = other.fd;
+    name = std::move(other.name);
+    other.fd = -1;
+    other.name.clear();
+    return *this;
+  }
+
+  ~TempFile() {
+    if (fd >= 0) {
+      unlink(name.c_str());
+      close(fd);
+    }
+  }
+
+  int fd;
 #endif // !defined(_WIN32)
 
-  const std::string name;
+  std::string name;
 };
 
 /// Attempts to return a temporary file or returns `nullopt` if an error
@@ -98,7 +115,7 @@ inline c10::optional<TempFile> try_make_tempfile(
 /// not be returned.
 inline TempFile make_tempfile(std::string name_prefix = "torch-file-") {
   if (auto tempfile = try_make_tempfile(std::move(name_prefix))) {
-    return *tempfile;
+    return std::move(*tempfile);
   }
   TORCH_CHECK(false, "Error generating temporary file: ", std::strerror(errno));
 }
